Fix the following bugs:
492: Video doesn't stop playing
- When the user stops playing, the video player actually sets the position of the media to 0 and then plays a very short segment of video so that a frame appears in the app window. The player is then supposed to stop playing. On Windows, setting the media position to a negative position doesn't behave the same way as it does on other platforms with the result that the JMF player failed to see the stop marker. The solution was to stop the video and then cue the video from 0.0 to some positive cue position. I also changed the behavior of skip forward and skip backward, to stop, reposition and start (if the video was playing). Previously the app simply set the position of the media, which works everywhere except Windows.

493: Video app doesn't idle when out of view
The video app wasn't handling the BOUNDS status change, so when an avatar left a region with a video player, the video player would continue to play, consuming cpu  and memory. Now when an avatar leaves a video app and the cell gets a BOUNDS message, it stops playing and also cancels any downloading of video files. The downside of this is that if the user is connected over a low bandwidth connection, the video cacheing process will have to restart when they return to the region containing the video app. In the future, it would be nice to continue the download in the background, or pause and resume it perhaps.

Lots of changes to debug messages for ease of log analysis.