Demoed at group meeting + bug fixes
Added a null check in setServerState - necessary because this is called before setLive, this fixes NPE adding MetaCompo to a new cell. Added missing code in for metadata caching (wasn't sending any mod messages but invalidate!).

Automatically adds wildcards around search terms. This currently reflects back into the search box, e.g. everything turns into "*xxxxx*". Left this way as an affordance and a reminder. In the future, if the search box adds a 'matches exactly' type of functionality, this could be removed.